Refactor analytics and order repository DTOs
- Updated `SalesAnalyticsDto` to use `decimal` for financial properties.
- Cleaned up `UniversityStudentCountDTO.cs` by removing unused directives and renaming the class.
- Adjusted `AveragePostPrice` assignment in `AnalyticsService.cs` to use `decimal` directly.
- Improved readability and performance in `OrderRepository.cs` with consistent method chaining and `AsNoTracking()`.
- Enhanced `CountTotalOrders` and simplified `AverageValueOfOrders` methods for better performance and clarity.

+++ Dentizone.Application/DTOs/Analytics/SalesAnalyticsDTO.cs
@@ -8,8 +8,8 @@ namespace Dentizone.Application.DTOs.Analytics
 {
     public class SalesAnalyticsDto
     {
-        public int TotalSalesRevenue { get; set; }
+        public decimal TotalSalesRevenue { get; set; }
         public int TotalsOrder { get; set; }
-        public int AveragePostPrice { get; set; }
+        public decimal AveragePostPrice { get; set; }
     }
 }
